+++ src/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -471,7 +471,7 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .end(".md"),
             ofFile("ext_mwb", "/icons/mysqlworkbench.png", "MySQL Workbench: *.mwb")
                 .end(".mwb"),
-            ofFile("ext_nsi", "/icons/nsis.png", "NSIS Nullsoft Scriptable Install System: *.nsi")
+            ofFile("ext_nsi", "/icons/nsis.svg", "NSIS Nullsoft Scriptable Install System: *.nsi")
                 .end(".nsi"),
             ofFile("ext_pdf", "/icons/pdf.png", "PDF: *.pdf")
                 .end(".pdf"),