+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * rework HTML5 files.
 * rework PID files.
 * improve FlyWay and Liquibase support: detect HSQLDB and IBM DB2 databases.
+* rework Angular files.
 
 ### 1.33.0 (2020/07/31)
 * support Robot Framework `.robot` files.