+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * rework WAR files.
 * rework TESTING files (PNG icon to SVG).
 * rework `.form` files (PNG icon to SVG).
+* rework NSIS files (PNG icon to SVG).
 
 Some icons come from the [Hiberbee Theme family](https://github.com/hiberbee/jetbrains-ide-theme) project.